slight reorganization, minor fixes of nothings, add tests
